requirements.txt
asgiref==3.8.1
certifi==2024.8.30
contourpy==1.3.0
crispy==0.8.0
cycler==0.12.1
distlib==0.3.8
dj-database-url==2.3.0
Django==5.1.1
django-allauth==65.0.1
django-crispy-forms==2.3
django-environ==0.11.2
django-heroku==0.3.1
djangorestframework==3.15.2
fabio==2024.9.0
filelock==3.16.1
fonttools==4.54.1
GDAL==3.6.2
geopandas==1.0.1
gis==0.2.1
greenlet==3.1.1
gunicorn==23.0.0
h5py==3.12.1
hdf5plugin==5.0.0
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.14.1
shapely==2.0.6
silx==2.1.1
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.5
whitenoise==6.7.0
xraydb==4.5.4
